requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==24.1.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   starlette
    #   watchfiles
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
coverage[toml]==7.6.1
    # via pytest-cov
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
execnet==2.1.1
    # via pytest-cache
fastapi==0.114.0
    # via -r requirements.in
flake8==5.0.4
    # via pytest-flake8
google-api-core[grpc]==2.19.2
    # via
    #   google-cloud-core
    #   google-cloud-firestore
google-auth==2.34.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-firestore
google-cloud-core==2.4.1
    # via google-cloud-firestore
google-cloud-firestore==2.20.0
    # via -r requirements.in
googleapis-common-protos==1.65.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.66.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.66.1
    # via google-api-core
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
idna==3.8
    # via
    #   anyio
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via -r requirements.in
markupsafe==2.1.5
    # via jinja2
mccabe==0.7.0
    # via flake8
mock==5.1.0
    # via -r requirements.in
packaging==24.1
    # via pytest
pep8==1.7.1
    # via pytest-pep8
pluggy==1.5.0
    # via pytest
proto-plus==1.24.0
    # via
    #   google-api-core
    #   google-cloud-firestore
protobuf==5.28.0
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pycodestyle==2.9.1
    # via flake8
pydantic==2.9.0
    # via fastapi
pydantic-core==2.23.2
    # via pydantic
pyflakes==2.5.0
    # via flake8
pytest==8.3.2
    # via
    #   -r requirements.in
    #   pytest-cache
    #   pytest-cov
    #   pytest-flake8
    #   pytest-pep8
pytest-cache==1.0
    # via pytest-pep8
pytest-cov==5.0.0
    # via -r requirements.in
pytest-flake8==1.2.2
    # via -r requirements.in
pytest-pep8==1.0.6
    # via -r requirements.in
python-dotenv==1.0.1
    # via uvicorn
pyyaml==6.0.2
    # via
    #   responses
    #   uvicorn
requests==2.32.3
    # via
    #   google-api-core
    #   responses
responses==0.25.3
    # via -r requirements.in
rsa==4.9
    # via google-auth
sniffio==1.3.1
    # via anyio
starlette==0.38.5
    # via fastapi
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   uvicorn
tzdata==2024.1
    # via pydantic
urllib3==2.2.2
    # via
    #   requests
    #   responses
uvicorn[standard]==0.30.6
    # via -r requirements.in
uvloop==0.20.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.0.1
    # via uvicorn
